add combination loss
